requirements.txt
elevenlabs
python-dotenv
groq
langchain
pydantic
langgraph
openai
Flask
python-dotenv
requests
PyGithub
PyMuPDF
pdfplumber